feat(dashboard): add reading page and card stats component
- added `reading.tsx` page under dashboard features
- created reusable `card-stats.tsx` component
- updated `app/dashboard/reading/page.tsx` to use new structure
